Moved create_swapchain to new file
